+++ hpc/nvidia/cuda/fakeCUDA/cuda.cpp
@@ -1,12 +1,27 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
+#ifdef HAVE_HPCC
+#include <hc_runtime.h>
+#include <hcc/hcc_internal.h>
+#include <cuda_adapter.h>
+#else
 #include <cuda.h>
+#endif
 #include "device.h"
 
 
 #ifdef __cplusplus
 extern "C" {
 #endif
+
+#ifdef HAVE_HPCC
+# define CUresult	hcError_t
+# undef CUDA_ERROR_INVALID_VALUE
+# define CUDA_ERROR_INVALID_VALUE	cudaErrorInvalidValue
+# undef CUDA_SUCCESS
+# define CUDA_SUCCESS	cudaSuccess
+#endif
+
 CUresult cuDeviceGet(CUdevice *device, int ordinal)
 {
 	if (ordinal < 0 || ordinal >= dev_count())